Handle race between createSession() and receipt of first message.
When the local side is the recipient of an incoming SIP call for
MSRP, and wants to wait until the call is answered before setting
up the MSRP session, there is an inherent race condition, because
in SIP the remote side knows about the successful call setup (OK)
before the local side (ACK), and therefore the local side may
receive the incoming MSRP TCP connection before it has had a chance
to configure MSRP with the new session, and as a result the remote
side would get a bogus "481 Session does not exist".

To fix this we add a hack to hang on to "orphan" incoming MSRP
requests for 500ms to give the local application a chance to setup
a matching session.